ml_project/requirements.txt
# local package
-e .

# external requirements
boto3
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
pandas
